pyproject.toml
[project]
name = "mcp-nt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.7.0",
    "pyntcore>=2025.3.2.3",
]
